requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
drf-yasg==1.21.10
gotrue==2.12.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jiter==0.10.0
numpy==2.3.1
openai==1.93.0
packaging==25.0
pgvector==0.4.1
pluggy==1.6.0
postgrest==1.1.1
psycopg==3.2.9
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
realtime==2.5.3
regex==2024.11.6
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
